+++ CHANGELOG.md
@@ -24,6 +24,7 @@ develop
     - Add dot mark to syntax names in the list to represent the state if style is customized.
     - Enable restoring modified syntax style directly from the list without opening the style editor.
 - Now, the current line number is drawn in bold font, and always drawn in vertical text mode.
+- Select whole text wrapped with quotation marks by double-clicking one of the quotation marks if it is already syntax-highlighted.
 - Keep text selection after inserting color code from the color code panel.
 - Update default fonts.
 - Add “description” field also to outline setting in syntax style editor.